Components Reorganized
Components removed form App.js and each put into their own file
Fixes to syntax made
Next step: creat a home screen and navigator
