+++ App.js
@@ -1,169 +1,12 @@
 import React from 'react';
 import { Alert, AppRegistry, Button, StyleSheet, Text, TextInput, View } from 'react-native';
 import { StackNavigator } from 'react-navigation';
-import Picker from 'react-native-picker';
 import { Dropdown } from 'react-native-material-dropdown';
-
-// export default class LogIn extends React.Component {
-//   // static navigationOptions = {
-//   //   title: 'Sign Up',
-//   // };
-//   constructor(props) {
-//     super(props);
-//     this.state = {text: ''};
-//   }
-//
-//   render() {
-//     return (
-//       <View>
-//         <View style={{padding: 10}}>
-//           <TextInput
-//             style={{height: 40}}
-//             placeholder="Username"
-//           />
-//           <Text style={{padding: 10, fontSize: 42}}>
-//           </Text>
-//         </View>
-//         <View style={{padding: 10}}>
-//           <TextInput
-//             style={{height: 40}}
-//             placeholder="Password"
-//           />
-//           <Text style={{padding: 10, fontSize: 42}}>
-//           </Text>
-//         </View>
-//         <Button
-//           onPress={() => { Alert.alert('Account Created!')}}
-//           title="LOG IN"
-//         />8
-//       </View>
-//     );
-//   }
-// }
-
-// export default class SignUp extends React.Component {
-//   // static navigationOptions = {
-//   //   title: 'Sign Up',
-//   // };
-//   constructor(props) {
-//     super(props);
-//     this.state = {text: ''};
-//   }
-//
-//   render() {
-//     return (
-//       <View>
-//         <View style={{padding: 10}}>
-//           <TextInput
-//             style={{height: 40}}
-//             placeholder="First Name"
-//           />
-//           <Text style={{padding: 10, fontSize: 42}}>
-//           </Text>
-//         </View>
-//         <View style={{padding: 10}}>
-//           <TextInput
-//             style={{height: 40}}
-//             placeholder="Last Name"
-//           />
-//           <Text style={{padding: 10, fontSize: 42}}>
-//           </Text>
-//         </View>
-//         <View style={{padding: 10}}>
-//           <TextInput
-//             style={{height: 40}}
-//             placeholder="Email"
-//           />
-//           <Text style={{padding: 10, fontSize: 42}}>
-//           </Text>
-//         </View>
-//         <View style={{padding: 10}}>
-//           <TextInput
-//             style={{height: 40}}
-//             placeholder="Password"
-//           />
-//           <Text style={{padding: 10, fontSize: 42}}>
-//           </Text>
-//         </View>
-//         <Button
-//           onPress={() => { Alert.alert('Account Created!')}}
-//           title="SIGN UP"
-//         />
-//       </View>
-//     );
-//   }
-// }
-
-export default class DebateSignUp extends React.Component {
-  // static navigationOptions = {
-  //   title: 'Sign Up',
-  // };
-  constructor(props) {
-    super(props);
-    this.state = {text: ''};
-  }
-
-  render() {
-    let pref = [{
-      value: 'Debate',
-    }, {
-      value: 'Judge',
-    }, {
-      value: 'Indifferent',
-    }];
-    let skill = [{
-      value: 'Novice',
-    }, {
-      value: 'Pro',
-    }, {
-      value: 'Worlds',
-    }];
-    return (
-      <View>
-        <View style={{padding: 10}}>
-          <TextInput
-            style={{height: 40}}
-            placeholder="First Name"
-          />
-          <Text style={{padding: 10, fontSize: 42}}>
-          </Text>
-        </View>
-        <View style={{padding: 10}}>
-          <TextInput
-            style={{height: 40}}
-            placeholder="Last Name"
-          />
-          <Text style={{padding: 10, fontSize: 42}}>
-          </Text>
-        </View>
-        <View style={{padding: 10}}>
-          <TextInput
-            style={{height: 40}}
-            placeholder="Partner Preference"
-          />
-          <Text style={{padding: 10, fontSize: 42}}>
-          </Text>
-        </View>
-        <Dropdown
-          label='Debate Preference'
-          data={pref}
-        />
-        <Dropdown
-          label='Skill Level'
-          data={skill}
-        />
-        <Button
-          onPress={() => { Alert.alert('Account Created!')}}
-          title="SIGN UP TO DEBATE"
-        />
-      </View>
-    );
-  }
-}
 
 // const App = StackNavigator({
 //   SignUp: { screen: SignUp },
-//   // LogIn: { screen: LogIn },
+//   LogIn: { screen: LogIn },
+//   DebateSignUp: { screen: DebateSignUp }
 // });
 
 // AppRegistry.registerComponent('App', () => App);
